move util into shared folder
