+++ src/app/core/patient.service.ts
@@ -3,7 +3,7 @@ import {Environment, EnvironmentToken} from "../environment";
 import {HttpClient, HttpErrorResponse} from "@angular/common/http";
 import {Observable, pipe} from "rxjs/Rx";
 import { map, catchError } from 'rxjs/operators'
-import {getMessageFromBackendError} from "../utils";
+import {getMessageFromBackendError} from "../../shared/utils";
 import {Patient} from "../../dto/patient";
 
 interface PatientResponse {
